refactor: Update Access-Control-Allow-Origin header in next.config
This commit updates the Access-Control-Allow-Origin header in the next.config file. The previous value of "*" has been changed to "*" to allow requests from any origin. This change allows for more flexible cross-origin resource sharing. Additionally, the commit adds the Access-Control-Allow-Methods and Access-Control-Allow-Headers headers to specify the allowed methods and headers for CORS requests.